requirements.txt
discord.py
better_profanity
cryptography
requests